requirements.txt
#pylint: skip-file

pylint
pytest
python-dotenv
requests